+++ src/arcpbf/mod.rs
@@ -203,7 +203,7 @@ fn decode_dense_nodes(
         let timestamp = timestamps[index] as i32 + last_timestamp;
         let timestamp = timestamp * date_granularity;
         last_timestamp = timestamp;
-        let timestamp = TimestampFormat::EpochNunber(timestamp as i64);
+        let timestamp = TimestampFormat::EpochNumber(timestamp as i64);
         assert!(uid_id < i32::MAX);
 
         results.push(ArcOSMObj::Node(ArcNode {
@@ -273,7 +273,7 @@ fn decode_ways(
         //let timestamp = timestamp * date_granularity;
         //last_timestamp = timestamp;
         //let timestamp = epoch_to_iso(timestamp);
-        let timestamp = TimestampFormat::EpochNunber(way.get_info().get_timestamp());
+        let timestamp = TimestampFormat::EpochNumber(way.get_info().get_timestamp());
 
         results.push(ArcOSMObj::Way(ArcWay {
             _id: id,
@@ -359,7 +359,7 @@ fn decode_relations(
         //let timestamp = timestamp * date_granularity;
         //last_timestamp = timestamp;
         //let timestamp = epoch_to_iso(timestamp);
-        let timestamp = TimestampFormat::EpochNunber(relation.get_info().get_timestamp());
+        let timestamp = TimestampFormat::EpochNumber(relation.get_info().get_timestamp());
 
         results.push(ArcOSMObj::Relation(ArcRelation {
             _id: id,